facts/requirements.txt
pinecone
boto3
python-dotenv
